+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Added `flapping` field to check history, along with `is_flapping_start` and
+  `is_flapping_end` event properties for use by filters.
+- Fixed bug where flapping would incorrectly end when `total_state_change` was below
+  `high_flap_threshold` instead of below `low_flap_threshold`.
 
 ### Fixed
 - ensure that check/check config have a non-empty command